moves formatting of diffs into util file
